added service implente, and some change in atributtes

+++ src/main/java/proyecto_tingeso_1/DTOS/BookingResponseDTO.java
@@ -22,7 +22,7 @@ public class BookingResponseDTO {
 
     private Integer originalPriceBooking;
     private Integer discountedPriceBooking;
-    private Integer discountPercentage;
+    private double discountPercentage;
     private String discountTypeBooking;
